[thinfs] Use Fuchsia-defined open flags; align with fs flags
Re-lands dfa53912377158b0248644d0f5e062fc1a430080

Also roll Zircon to pull in the open flag changes.

+++ public/rust/crates/fdio/src/rio.rs
@@ -149,7 +149,7 @@ impl Message {
     }
 
     pub fn is_pipelined(&self) -> bool {
-        self.arg() & O_PIPELINE != 0
+        (self.arg() as u32) & O_PIPELINE != 0
     }
 }
